+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   * removed `--report` flag (it is always the default),
     added `--verbose` flag to display full JSON data
   * removed `--scores` flag (it is always-on by default)
+  * removed `--verbosity` option (the output data is always the same)
 * Removed:
   * `bin/bulk.dart`
   * `ComponentNames`
@@ -25,6 +26,7 @@
   * `Summary.platform`
   * `Summary.stats`
   * `Summary.suggestions`
+  * `Verbosity`
 
 ## 0.13.16